+++ flow-typed/mountain.js
@@ -37,29 +37,35 @@ declare module '@prostory/mountain' {
   declare type ServerOptions = {
     cert: Buffer,
     key: Buffer,
-    rootDir: string | URL,
     timeout?: number,
   }
 
-  declare type RouteOptions = {
+  declare type Route = {
     path: string | RegExp,
     method: string,
     notFound?: boolean,
     handle: (request: Http2Request, response: Http2Response) => void,
   }
 
-  declare export class Route {
-    constructor(options: RouteOptions): Route;
-
-    get path(): string;
-    get method(): string;
-    get isForNotFound(): boolean;
-
-    handle(request: Http2Request, response: Http2Response): void;
+  declare type ClientOptions = {
+    url: string,
+    maxSessionMemory?: number,
+    settings?: Http2Settings,
   }
 
-  declare export class Handler {
-    constructor(routes: Route[]): Handler;
+  declare type RequestOptions = {
+    onResponse: (headers: Http2Headers) => void,
+    onData: (chunk: Buffer) => void,
+    onEnd: () => void,
+    endStream?: boolean,
+    exclusive?: boolean,
+    parent?: number,
+    weight?: number,
+    waitForTrailers?: boolean,
+  }
+
+  declare export class Router {
+    constructor(routes: Route[]): Router;
 
     set(): (request: Http2Request, response: Http2Response) => void;
   }
@@ -76,6 +82,17 @@ declare module '@prostory/mountain' {
     onError(listener: (error: Error) => void): void;
     setTimeout(milliseconds?: number, callback?: () => void): void;
     listen(port: number, host: ?string, listener?: () => void): void;
+    close(callback?: () => void): void;
+  }
+
+  declare export class Client {
+    constructor(options: ClientOptions): Client;
+
+    onAltsvc(
+      listener: (alt: string, origin: string, streamId: number) => void
+    ): void;
+    onOrigin(listener: (origins: string[]) => void): void;
+    request(headers: Http2Headers, options: RequestOptions): void;
     close(callback?: () => void): void;
   }
 }