+++ src/errors/fs_error.mjs
@@ -1,0 +1,8 @@
+// @flow
+
+export class FsError extends Error {
+  constructor(message: string) {
+    super(message)
+    this.name = 'FsError'
+  }
+}