+++ src/router.mjs
@@ -1,12 +1,19 @@
 // @flow
 
-import { sendError } from '../send.mjs'
+import { sendError } from './send.mjs'
+import { constants } from 'http2'
 
-import type { Http2Request, Http2Response } from '../server.mjs'
-import type Route from './route.mjs'
+import type { Http2Request, Http2Response } from './server.mjs'
 
-export default class Handler {
-  +_routes: Route[]
+type Route = {|
+  path: string | RegExp,
+  method: string,
+  notFound?: boolean,
+  handle: (request: Http2Request, response: Http2Response) => void,
+|}
+
+export class Router {
+  _routes: Route[]
 
   constructor(routes: Route[]) {
     this._routes = routes
@@ -17,7 +24,7 @@ export default class Handler {
       const path = request.headers[':path']
       const method = request.headers[':method']
 
-      const route = this._routes.find(r => {
+      const route = this._routes.find((r) => {
         const pathRegExp =
           typeof r.path === 'string' ? new RegExp(`^${r.path}$`) : r.path
 
@@ -30,14 +37,14 @@ export default class Handler {
       if (route) {
         route.handle(request, response)
       } else {
-        const notFoundRoute = this._routes.find(route => route.isForNotFound)
+        const notFoundRoute = this._routes.find((route) => route.notFound)
 
         if (notFoundRoute) {
           notFoundRoute.handle(request, response)
         } else {
           sendError(request.stream, {
             status: 404,
-            reason: `No route that match specified path: ${path}`
+            reason: `No route that match specified path: ${path}`,
           })
         }
       }